Refactor MCTSMetrics integration and improve modularity.
Moved MCTSMetrics to utils and updated imports for consistency. Added optional MCTSMetrics parameter to support custom metrics tracking, improving flexibility across self-play and experiment workflows.
